% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{annotate.bedGraph.tracks}
\alias{annotate.bedGraph.tracks}
\title{return annotation track of hotspots with alignability/uniqueness scores (UCSC)}
\usage{
annotate.bedGraph.tracks(df, bedgraph)
}
\arguments{
\item{df}{numeric first argument}

\item{bedgraph}{numeric second argument}
}
\value{
numeric (aa$mu_position[k]/mu_prot)
}
\description{
return annotation track of hotspots with alignability/uniqueness scores (UCSC)
}
