% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{annotate.center.bias}
\alias{annotate.center.bias}
\title{return TRUE|FALSE of hotspots with putative mutation calling center bias}
\usage{
annotate.center.bias(sig, maf)
}
\arguments{
\item{sig}{numeric first argument}

\item{maf}{numeric second argument}
}
\value{
return TRUE|FALSE of hotspots with putative mutation calling center bias
}
\description{
return TRUE|FALSE of hotspots with putative mutation calling center bias
}
