% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{annotate.entropy}
\alias{annotate.entropy}
\title{return calculcated entropy around the site of each hotspot}
\usage{
annotate.entropy(sig)
}
\arguments{
\item{sig}{numeric first argument}
}
\value{
numeric (aa$mu_position[k]/mu_prot)
}
\description{
return calculcated entropy around the site of each hotspot
}
