% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{annotate.filtering.judgement}
\alias{annotate.filtering.judgement}
\title{return filtering judgement based on annotated metrics}
\usage{
annotate.filtering.judgement(sig)
}
\arguments{
\item{sig}{numeric first argument}
}
\value{
return filtering judgement based on annotated metrics
}
\description{
return filtering judgement based on annotated metrics
}
