% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{annotate.homopolymers}
\alias{annotate.homopolymers}
\title{homopolymer region metrics: TRUE|FALSE if it is a repeat region, the repeat nucleotide(s), and length of repeat region}
\usage{
annotate.homopolymers(sig, homopolymerbed)
}
\arguments{
\item{sig}{numeric first argument}

\item{homopolymerbed}{numeric second argument}
}
\value{
homopolymer region metrics: TRUE|FALSE if it is a repeat region, the repeat nucleotide(s), and length of repeat region
}
\description{
homopolymer region metrics: TRUE|FALSE if it is a repeat region, the repeat nucleotide(s), and length of repeat region
}
