% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{annotate.true.positives}
\alias{annotate.true.positives}
\title{return TRUE|FALSE of whether hotspot is a true positive}
\usage{
annotate.true.positives(sig, dmp)
}
\arguments{
\item{sig}{numeric first argument}

\item{dmp}{numeric second argument}
}
\value{
numeric TRUE|FALSE of whether hotspot is a true positive
}
\description{
return TRUE|FALSE of whether hotspot is a true positive
}
