% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{any.neighbors}
\alias{any.neighbors}
\title{returns TRUE if there are other positions that are 1 bp away}
\usage{
\method{any}{neighbors}(df)
}
\arguments{
\item{df}{numeric first argument}
}
\value{
returns TRUE if there are other positions that are 1 bp away
}
\description{
returns TRUE if there are other positions that are 1 bp away
}
