% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{binom.test_snp}
\alias{binom.test_snp}
\title{perform binomial test for all mutations observed in gene}
\usage{
binom.test_snp(gene)
}
\arguments{
\item{gene}{numeric first argument}
}
\value{
perform binomial test for all mutations observed in gene
}
\description{
perform binomial test for all mutations observed in gene
}
