% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{combine}
\alias{combine}
\title{returns the table and column names of table into a single string}
\usage{
combine(tb, sep = ":")
}
\arguments{
\item{tb}{numeric first argument}

\item{sep}{character string second argument}
}
\value{
returns the table and column names of table into a single string
}
\description{
returns the table and column names of table into a single string
}
