% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{convert.to.decimal}
\alias{convert.to.decimal}
\title{return perentage into decimal}
\usage{
convert.to.decimal(num)
}
\arguments{
\item{num}{numeric first argument}
}
\value{
numeric (aa$mu_position[k]/mu_prot)
}
\description{
return perentage into decimal
}
