% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{find.splice.neighbors}
\alias{find.splice.neighbors}
\title{returns the mutability of the codon given the gene}
\usage{
find.splice.neighbors(pos, df)
}
\arguments{
\item{pos}{numeric first argument}

\item{df}{numeric second argument}
}
\value{
returns other positions that are 1 bp away from pos
}
\description{
returns other positions that are 1 bp away from pos
}
