% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{get.af.rank}
\alias{get.af.rank}
\title{return all the VAF ranks of the mutation in all samples with the mutation}
\usage{
get.af.rank(pos, maf)
}
\arguments{
\item{pos}{numeric first argument}

\item{maf}{numeric second argument}
}
\value{
details
}
\description{
return all the VAF ranks of the mutation in all samples with the mutation
}
