% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{get.all.tri}
\alias{get.all.tri}
\title{returns the mutability of the codon given the gene}
\usage{
get.all.tri(pos, maf)
}
\arguments{
\item{pos}{numeric first argument}

\item{maf}{numeric second argument}
}
\value{
return the counts and types of tri-nucleotide contexts
}
\description{
returns the mutability of the codon given the gene
}
