% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{get.alpha}
\alias{get.alpha}
\title{returns the mutability of the codon given the gene}
\usage{
get.alpha(k, aa, mu_prot)
}
\arguments{
\item{k}{numeric first argument}

\item{aa}{numeric second argument}

\item{mu_prot}{numeric third argument}
}
\value{
numeric (aa$mu_position[k]/mu_prot)
}
\description{
returns the mutability of the codon given the gene
}
