% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{get.cancer.type}
\alias{get.cancer.type}
\title{returns the number of cancer types with the mutation}
\usage{
get.cancer.type(position, mf, d = d)
}
\arguments{
\item{position}{numeric first argument}

\item{mf}{numeric second argument}

\item{d}{numeric third argument}
}
\value{
returns the number of cancer types with the mutation
}
\description{
returns the number of cancer types with the mutation
}
