% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{get.ccf}
\alias{get.ccf}
\title{return the cancer cell fractions (CCF) of the mutation}
\usage{
get.ccf(pos, maf)
}
\arguments{
\item{pos}{numeric first argument}

\item{maf}{numeric second argument}
}
\value{
details
}
\description{
return the cancer cell fractions (CCF) of the mutation
}
