% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{get.mu.score}
\alias{get.mu.score}
\title{return the weighted average tri-nucleotide mutability of the codon}
\usage{
get.mu.score(pos, maf, mu)
}
\arguments{
\item{pos}{numeric first argument}

\item{maf}{numeric second argument}

\item{mu}{numeric third argument}
}
\value{
return the weighted average tri-nucleotide mutability of the codon
}
\description{
return the weighted average tri-nucleotide mutability of the codon
}
