% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{get.neighbors}
\alias{get.neighbors}
\title{returns the first occurance of a neighbor}
\usage{
get.neighbors(df)
}
\arguments{
\item{df}{numeric first argument}
}
\value{
details
}
\description{
returns the first occurance of a neighbor
}
