% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{get.probability}
\alias{get.probability}
\title{returns baseline expected probability given gene length, gene mutation burden, and total samples}
\usage{
get.probability(gene, aa, total.muts, total.samples, aa.length)
}
\arguments{
\item{gene}{numeric first argument}

\item{aa}{numeric second argument}

\item{total.muts}{numeric third argument}

\item{total.samples}{numeric fourth argument}

\item{aa.length}{numeric fifth argument}
}
\value{
numeric (1/aa.length)*(total.muts/total.samples)
}
\description{
returns baseline expected probability given gene length, gene mutation burden, and total samples
}
