% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{get.pvalues}
\alias{get.pvalues}
\title{returns log10 p-value of the binomial distribution}
\usage{
get.pvalues(k, total.muts, aa.length, mu_prot, aa, gene, min_prob,
  total.samples = TOTAL_SAMPLES)
}
\arguments{
\item{k}{numeric first argument}

\item{total.muts}{numeric second argument}

\item{mu_prot}{numeric fourth argument}

\item{aa}{numeric fifth argument}

\item{gene}{numeric sixth argument}

\item{min_prob}{numeric seventh argument}

\item{total.samples}{numeric seventh argument}

\item{as.length}{numeric third argument}
}
\value{
pbinom(as.numeric(as.character(aa$count[k]))-1,size=total.samples,prob=prob,lower.tail=FALSE,log.p=TRUE)/log(10,exp(1))
}
\description{
returns log10 p-value of the binomial distribution
}
