% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{get.reference.aa}
\alias{get.reference.aa}
\title{returns the reference amino acid}
\usage{
get.reference.aa(pos, maf)
}
\arguments{
\item{pos}{numeric first argument}

\item{maf}{numeric second argument}
}
\value{
returns the reference amino acid
}
\description{
returns the reference amino acid
}
