% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{get.reference.tri}
\alias{get.reference.tri}
\title{returns the count of tri-nucleotide contexts of the codon}
\usage{
get.reference.tri(pos, maf)
}
\arguments{
\item{pos}{numeric first argument}

\item{maf}{numeric second argument}
}
\value{
returns the count of tri-nucleotide contexts of the codon
}
\description{
returns the count of tri-nucleotide contexts of the codon
}
