% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{get.repeats}
\alias{get.repeats}
\title{homopolyer region, if any, of a given hotspot}
\usage{
get.repeats(i, sig, rr)
}
\arguments{
\item{i}{numeric first argument}

\item{sig}{numeric second argument}

\item{rr}{third argument}
}
\value{
homopolyer region, if any, of a given hotspot
}
\description{
homopolyer region, if any, of a given hotspot
}
