% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{get.rsid}
\alias{get.rsid}
\title{returns dbSNP/1000G/NHLBI ids, if any}
\usage{
get.rsid(pos, maf)
}
\arguments{
\item{pos}{numeric first argument}

\item{maf}{numeric second argument}
}
\value{
returns dbSNP/1000G/NHLBI ids, if any
}
\description{
returns dbSNP/1000G/NHLBI ids, if any
}
