% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{get.sample.rank}
\alias{get.sample.rank}
\title{returns the mutability of the codon given the gene}
\usage{
get.sample.rank(sample, gene, pos, maf)
}
\arguments{
\item{sample}{numeric first argument}

\item{gene}{numeric second argument}

\item{pos}{numeric third argument}

\item{maf}{numeric fourth argument}
}
\value{
return the variant allele frequency (VAF) ranks of the mutation in the sample
}
\description{
return the variant allele frequency (VAF) ranks of the mutation in the sample
}
