% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{get.variant.aa}
\alias{get.variant.aa}
\title{returns the number and types of mutant residues of the codon}
\usage{
get.variant.aa(position, mf)
}
\arguments{
\item{position}{numeric first argument}

\item{mf}{numeric second argument}
}
\value{
numeric (aa$mu_position[k]/mu_prot)
}
\description{
returns the number and types of mutant residues of the codon
}
