% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{mut.center.breakdown}
\alias{mut.center.breakdown}
\title{return mutation calling centers (TCGA) of a given hotspot}
\usage{
mut.center.breakdown(i, sig, maf)
}
\arguments{
\item{i}{numeric first argument}

\item{sig}{numeric second argument}

\item{maf}{numeric third argument}
}
\value{
return mutation calling centers (TCGA) of a given hotspot
}
\description{
return mutation calling centers (TCGA) of a given hotspot
}
