% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_game.R
\name{new_game}
\alias{new_game}
\title{Start a New Game}
\usage{
new_game()
}
\value{
list with the current spot, number of chances the player has left, and seed
}
\description{
Start a New Game
}
\details{
We want to start a new game of Hot Spots.  This function begins a new game
by setting the player's current spot to spot 1 (the starting spot).
Displays the board with the player at spot 1.  Selects a random number
as the seed for the current game to use in the move_to function.  This is
used to select and retain the hot spots for the current game.  Starts the
player with three chances.
}
\examples{
new_game()
}
\author{
S. Ser
}
