% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{prepmaf}
\alias{prepmaf}
\title{prep the MAF}
\usage{
prepmaf(maf, expressiontb)
}
\arguments{
\item{maf}{character string Path to MAF}

\item{expressiontb}{numeric second argument}
}
\value{
MAF
}
\description{
prep the MAF (revise)
}
