% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{remove.snps}
\alias{remove.snps}
\title{deprecated germline SNP filtering based on 1000/NHLBI}
\usage{
remove.snps(maf)
}
\arguments{
\item{maf}{character string Path to MAF}
}
\value{
numeric (aa$mu_position[k]/mu_prot)
}
\description{
deprecated germline SNP filtering based on 1000/NHLBI
putative germline SNPs are filtered based on ExAC with minor AF > 0.06%
ExAC r0.2 has been preloaded
}
