% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{remove.unexpressed.genes}
\alias{remove.unexpressed.genes}
\title{details}
\usage{
remove.unexpressed.genes(i, maf, express)
}
\arguments{
\item{i}{numeric first argument}

\item{maf}{numeric second argument}

\item{express}{numeric third argument}
}
\value{
details
}
\description{
details
}
