% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{remove.unexpressed.mutations}
\alias{remove.unexpressed.mutations}
\title{remove unexpressed mutations}
\usage{
remove.unexpressed.mutations(maf, expressiontb)
}
\arguments{
\item{maf}{character string Path to MAF}

\item{expressiontb}{numeric second argument}
}
\value{
MAF
}
\description{
remove unexpressed mutations (revise)
}
