% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{split.double}
\alias{split.double}
\title{details}
\usage{
\method{split}{double}(ind, df)
}
\arguments{
\item{ind}{numeric first argument}

\item{df}{numeric second argument}
}
\value{
numeric (aa$mu_position[k]/mu_prot)
}
\description{
details
}
