% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lchs.R
\name{Lchs}
\alias{Lchs}
\title{Lorenz curve with hot spot cutoff}
\usage{
Lchs(x, ...)
}
\arguments{
\item{x}{\code{"hotspots"} object}

\item{...}{further plotting parameters to pass to \code{\link[ineq]{plot.Lc}}}
}
\description{
Plot a Lorenz curve with a hot spot cutoff on it.
}
\details{
Uses the function \code{\link[ineq]{plot.Lc}} from the \code{ineq} package to plot a Lorenz curve based on the data in a \code{hotspots} object. The location of the hot spot cutoff on the Lorenz curve is then drawn as a filled black circle.
}
\examples{
Lchs(hotspots(rlnorm(100)))

}
\seealso{
\code{\link{hotspots}}, \code{\link[ineq]{Lc}}, \code{\link[ineq]{plot.Lc}}
}
\author{
Anthony Darrouzet-Nardi
}
