% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.hotspots.R
\name{plot.hotspots}
\alias{plot.hotspots}
\title{Plotting hot spot and outlier cutoffs}
\usage{
\method{plot}{hotspots}(x, pch = par("pch"), ...)
}
\arguments{
\item{x}{\code{"hotspots"} object}

\item{pch}{plotting character. See \code{\link{par}}}

\item{...}{further plotting parameters to pass to \code{\link[lattice]{densityplot}}}
}
\value{
An object of class \code{"trellis"}.
}
\description{
\code{plot} method for class \code{"hotspots"}.
}
\details{
Uses the function \code{\link[lattice]{densityplot}} from the \code{lattice} package to show the distribution of the data and the position
of the positive and/or negative hot spot or outlier cutoffs.
}
\examples{
#both tails on skewed data
rln100pn <- hotspots(c(rlnorm(50),rlnorm(50)*-1),tail = "both")
plot(rln100pn)

#modify graphical parameters
plot(rln100pn, pch = 16, cex = 1.5)

}
\seealso{
\code{\link{hotspots}}, \code{\link{summary.hotspots}}, \code{\link[lattice]{densityplot}}
}
\author{
Anthony Darrouzet-Nardi
}
